basic simulator set up working, descent glide checked
CX = 0.01
CL = 2pi*alpha (where alpha is in radians)

The initial climb angle can be specified, and is current set to 10 degrees. We can see where the aircraft settles - at alpha = 10.5 degrees, giving CL = 1.152 and CX = 0.01. Theoretically this should have a glide descent angle of 0.497 degrees, and simulink says that in 25 seconds the glide angle settles, and limits to -0.502 degrees. So pretty good